move tests to common folder
