+++ components/google_search_console/common/tests/mockery-dollar.mjs
